Add files via upload
This is the most proper way to connect the battery to LED using a voltage regulator. We will use the parallel connection. This is because it's more efficient especially if there is more than one LED. The current will split into LEDs and if one is damaged, others will continue to work.